feat!: Offer simplified collapse defaults in the API
BREAKING CHANGE: This changes the interface used for the collapse options and shortens it to collapse.  Otherwise the options remain the same.  The "direction" attribute has also been shortened to match the HTML attribute that it draws from.  The API is otherwise the same.
